unrevert:
Merge pull request #169 in HOT/datarouter from mcorgan/counter-adapter/DATAROUTER-169 to master

* commit '5117aff176946c8bbf03fb5e2d5079eac1838605': (34 commits)
  add a few missing scanning counters
  have ClientTypes return Physical CallsiteAdapters
  add toString method to Counter and CallsiteAdapter
  hint at removing Node interface methods
  rename AdapterNodes to CallsiteAdapters
  remove some (hopefully) temporary uses of node.getPhysicalNodeIfApplicable()
  wrap all PhysicalNodes in a PhysicalCounterAdapter
  these nodes need to implement a Physical interface
  ClientType can return PhysicalNode
  add Physical CounterAdapters
  add some new interface combinations
  add physical CounterAdapters
  remove some casting for NodeId constructor
  cleanup Partitions.java
  remove unused methods
  switch DataRouter counters to lowercase-r Datarouter
  replace casts with getPhysicalNodeIfApplicable from CounterAdapters
  register missing ClientId.  not sure how this was passing before
  test for DefaultClientTypes strings to verify classes haven't moved
  remove exception catch thanks to TestNG
  ... (reverted from commit 3e1fb45afcc339a7dd9184136fca794936a08e39) (reverted from commit d37afecbeddf3af296b67bfddbbe8a0bb2bbc1b6)

+++ src/main/java/com/hotpads/datarouter/test/node/basic/manyfield/BaseManyFieldIntegrationTests.java
@@ -52,11 +52,8 @@ public abstract class BaseManyFieldIntegrationTests{
 	}
 
 	private void resetTable(){
-		try{
+		if(!isMemcached()){
 			mapNode.deleteAll(null);
-		}catch(UnsupportedOperationException e){
-			//swallow memcached unsupported op.  should probably take deleteAll out of the MapStorage interface
-			//too bad i can't call the isMemcached method from this static method
 		}
 	}
 
@@ -508,7 +505,8 @@ public abstract class BaseManyFieldIntegrationTests{
 	public void testIncrement(){
 		if(!isHBase()){ return; }
 		@SuppressWarnings("unchecked")
-		HBaseNode<ManyFieldBeanKey, ManyFieldBean, ?> hBaseNode = (HBaseNode<ManyFieldBeanKey,ManyFieldBean,?>)mapNode;
+		HBaseNode<ManyFieldBeanKey, ManyFieldBean, ?> hBaseNode = (HBaseNode<ManyFieldBeanKey,ManyFieldBean,?>)mapNode
+				.getPhysicalNodeIfApplicable();
 		ManyFieldBean bean = new ManyFieldBean();
 
 		//increment by 3
